feat: ajusta o component header
