+++ src/frontend/src/pages/fitaMedicamentos.jsx
@@ -1,7 +1,6 @@
-// LoginPage.jsx
-import React, { useState } from "react";
+import React from "react";
 import Table from "../components/Table";
-import Search from "../components/Search";
+import Header from "../components/Header"; // Agora só usamos o Header
 import "../styles/fitaMedicamentos.css";
 
 const dataAFazer = [
@@ -22,20 +21,15 @@ const dataProntas = [
     { nome: "Fita 1", descricao: "Supporting line text lorem ipsum dolor sit amet, consectetur." },
 ];
 
-
 export default function FitaMedicamentos() {
     return (
         <>  
-            <div className="header">
-            <h1>Fita de Medicamentos</h1>
-            <Search></Search>
-            </div>
-            <div className="conteudo" >
-            <Table title="A fazer" data={dataAFazer} maxItems={2} />
-            <Table title="Em progresso" data={dataEmProgresso} maxItems={1} />
-            <Table title="Prontas" data={dataProntas} maxItems={2} />
+            <Header title="Fita de Medicamentos" />
+            <div className="conteudo">
+                <Table title="A fazer" data={dataAFazer} maxItems={2} />
+                <Table title="Em progresso" data={dataEmProgresso} maxItems={1} />
+                <Table title="Prontas" data={dataProntas} maxItems={2} />
             </div>
         </>
-
-    )
+    );
 }
