+++ src/frontend/src/styles/Header.css
@@ -1,3 +1,17 @@
+.header {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding: 18px 0;
+    font-family: "Roboto", sans-serif;
+}
+
+h1 {
+    font-size: 26px;
+    font-weight: 400;
+    margin: 0;
+}
+
 .search-container {
     display: flex;
     align-items: center;
