+++ src/frontend/src/styles/fitaMedicamentos.css
@@ -9,20 +9,6 @@ html {
     display: flex;
     align-items: start;
     justify-content: center;
-    font-family: "Roboto", sans-serif;
-}
-
-.header {
-    display: flex;
-    justify-content: space-between;
-    padding: 18px 0;
-    align-items: center;
-}
-
-h1 {
-    font-size: 26px;
-    font-weight: 400;
-    margin: 0;
 }
 
 .conteudo {
@@ -30,5 +16,5 @@ h1 {
     align-items: center;
     flex-direction: column;
     gap: 25px;
-    margin: 30px 0;
+    margin: 15px 0 30px 0;
 }
